@W-16589746: [Android] REST wrappers for select SFAP APIs (Move sfap_api REST wrappers to REST package)
